Вернул в регистрацию в AuthController
