+++ src/main/java/ru/skypro/homework/controller/AuthController.java
@@ -30,4 +30,13 @@ public class AuthController {
         }
     }
 
+    @PostMapping("/register")
+    public ResponseEntity<?> register(@RequestBody Register register) {
+        if (authService.register(register)) {
+            log.info("Создан " + register);
+            return ResponseEntity.status(HttpStatus.CREATED).build();
+        } else {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+        }
+    }
 }
