--- src/main/java/ru/skypro/homework/controller/RegController.java
@@ -1,29 +1,0 @@
-package ru.skypro.homework.controller;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
-import ru.skypro.homework.dto.Register;
-import ru.skypro.homework.service.AuthService;
-
-@Slf4j
-@CrossOrigin(value = "http://localhost:3000")
-@RestController
-@RequiredArgsConstructor
-public class RegController {
-    private final AuthService authService;
-    @PostMapping("/register")
-    public ResponseEntity<?> register(@RequestBody Register register) {
-        if (authService.register(register)) {
-            log.info("Создан " + register);
-            return ResponseEntity.status(HttpStatus.CREATED).build();
-        } else {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
-        }
-    }
-}
